Removed old carousel, updated with new one.
Implemented a jQuery solution by ken wheeler, called the slick carousel.
The responsive design attracted me to this solution, however, I do not
like the defaults provided. I may have to go against DRY principles
because this library requires the use of <img> tags in the html
document. This is inconvenient because there is not an immediate
solution to me to properly style these images, and images a part of the
content may be different sizes. One solution would be to unify all image
sizes (which is probably best practice). However, due to the nature of
this site, the full-viewport height design is causing problems with the
formatting. The goal for this site is to first and foremost be
responsive, and viewable on any device, so sacrificing DRY methods here
is looking to be an unwanted but necessary tradeoff to keep things
intuitive. Tried implementing a solution with javascript, but the
complexity of callbacks needed from the library outweighed its
advantages.